fixes from first attempt at 'production' deployment
we need to be able to configure both the base url for the actual data
objects and the uri prefix that is stored for the data objects in the
triple store this might or might not be the same as the actual url on
which the service is deployed, both because it might be behind a proxy
and we want any CITE collection data stored in the triple store to
always reference stable urls untied to any specific implementation of a
CITE service and not be tightly coupled to the specific service that was
used to create them. In the long run, we should reference the service that
was used to create them as provenance data but that might end up being
something more complex that a simple source url reference and even then it
would also require stable uri identifiers untied to the deployment
environment.

another fix is just to make the error handling for the triple store inter-
actions a little more robust. We should catch and report errors if these
happen for any reason, not just the specific exceptions thrown by
the JackRDF gem. As it was coded before we got success return codes even
when the data wasn't actually stored.