adding a temporary solution for serving referenced images
we need a way, given the URN for an item in an image cite collection to
return the source image for the collection item. Ideally this probably should
not be done via JackSON but a server sitting on top of it - this is a temporary
quick solution to enable deployment on Perseids.